(LTH-149) Handle errors if child has exited before all writing finished
In testing, we periodically see `failed to run: stdin pipe i/o failed:
Broken pipe (32)` on AIX. My hypothesis is that this occurs because the
child process exits before we've made a final attempt to write 0 bytes
to it and close the pipe. Catch EPIPE when writing 0 bytes and consider
the pipe closed.